Feat: 지난 월간 통계 조회 기능

+++ src/main/java/joon/homework/dto/stats/response/NameAndCount.java
@@ -10,10 +10,10 @@ public class NameAndCount {
 
     private Long userId;
     private String name;
-    private Long count;
+    private int count;
 
     @Builder
-    public NameAndCount(Long userId, String name, Long count) {
+    public NameAndCount(Long userId, String name, int count) {
         this.userId = userId;
         this.name = name;
         this.count = count;
